remove unneccessary solvers (hi thomas)
