+++ include/common/globalconstants.hpp
@@ -124,31 +124,21 @@ inline std::map<std::string, KERNEL_NAME> Kernel_Map{ { "ISOTROPIC", KERNEL_Isot
 // Solver name
 enum SOLVER_NAME {
     SN_SOLVER,
-    CSD_SN_SOLVER,
-    CSD_SN_NOTRAFO_SOLVER,
-    CSD_SN_FOKKERPLANCK_SOLVER,
-    CSD_SN_FOKKERPLANCK_TRAFO_SOLVER,
-    CSD_SN_FOKKERPLANCK_TRAFO_SOLVER_2D,
-    CSD_SN_FOKKERPLANCK_TRAFO_SH_SOLVER_2D,
-    CSD_PN_SOLVER,
-    CSD_MN_SOLVER,
     PN_SOLVER,
     MN_SOLVER,
-    MN_SOLVER_NORMALIZED
+    MN_SOLVER_NORMALIZED,
+    CSD_SN_SOLVER,
+    CSD_PN_SOLVER,
+    CSD_MN_SOLVER,
 };
 
 inline std::map<std::string, SOLVER_NAME> Solver_Map{ { "SN_SOLVER", SN_SOLVER },
-                                                      { "CSD_SN_SOLVER", CSD_SN_SOLVER },
-                                                      { "CSD_SN_NOTRAFO_SOLVER", CSD_SN_NOTRAFO_SOLVER },
-                                                      { "CSD_SN_FOKKERPLANCK_SOLVER", CSD_SN_FOKKERPLANCK_SOLVER },
-                                                      { "CSD_SN_FOKKERPLANCK_TRAFO_SOLVER", CSD_SN_FOKKERPLANCK_TRAFO_SOLVER },
-                                                      { "CSD_SN_FOKKERPLANCK_TRAFO_SOLVER_2D", CSD_SN_FOKKERPLANCK_TRAFO_SOLVER_2D },
-                                                      { "CSD_SN_FOKKERPLANCK_TRAFO_SH_SOLVER_2D", CSD_SN_FOKKERPLANCK_TRAFO_SH_SOLVER_2D },
-                                                      { "CSD_PN", CSD_PN_SOLVER },
-                                                      { "CSD_MN", CSD_MN_SOLVER },
                                                       { "PN_SOLVER", PN_SOLVER },
                                                       { "MN_SOLVER", MN_SOLVER },
-                                                      { "MN_SOLVER_NORMALIZED", MN_SOLVER_NORMALIZED } };
+                                                      { "MN_SOLVER_NORMALIZED", MN_SOLVER_NORMALIZED },
+                                                      { "CSD_SN_SOLVER", CSD_SN_SOLVER },
+                                                      { "CSD_PN", CSD_PN_SOLVER },
+                                                      { "CSD_MN", CSD_MN_SOLVER } };
 
 // Entropy functional
 enum ENTROPY_NAME { QUADRATIC, MAXWELL_BOLTZMANN, BOSE_EINSTEIN, FERMI_DIRAC };
