+++ include/solvers/csdsnsolver.hpp
@@ -1,22 +1,38 @@
 #ifndef CSDSNSOLVER_H
 #define CSDSNSOLVER_H
 
+// externals
+#include "spdlog/spdlog.h"
+#include <mpi.h>
+
+#include "common/config.hpp"
+#include "common/io.hpp"
+#include "fluxes/numericalflux.hpp"
+#include "kernels/scatteringkernelbase.hpp"
+#include "problems/problembase.hpp"
+#include "quadratures/quadraturebase.hpp"
 #include "solvers/snsolver.hpp"
+#include "velocitybasis/sphericalharmonics.hpp"
+
+class Physics;
 
 class CSDSNSolver : public SNSolver
 {
   private:
     std::vector<double> _dose; /*!< @brief TODO */
 
-    // Physics acess
-    Vector _angle; /*!< @brief angles for SN */
+    Matrix _O;    // Transformation Matrix from Moments to Ordinates
+    Matrix _M;    // Transformation Matrix from Ordinates to Moments
 
-    std::vector<Matrix> _sigmaSE; /*!<  @brief scattering cross section for all energies*/
-    Vector _sigmaTE;              /*!<  @brief total cross section for all energies*/
+    VectorVector _quadPointsSphere;
+    Vector _eTrafo;
+
+    // Helper variables
+    unsigned _polyDegreeBasis;
 
   public:
     /**
-     * @brief CSDSNSolver constructor
+     * @brief CSDSolverTrafoFP2D constructor
      * @param settings stores all needed information
      */
     CSDSNSolver( Config* settings );
@@ -24,13 +40,24 @@ class CSDSNSolver : public SNSolver
     virtual ~CSDSNSolver() {}
 
     /**
-     * @brief Solve functions runs main time loop
+     * @brief Output solution to VTK file (LEGACY. But not remove, to compare)
      */
-    void Solve() override;
+    // virtual void Save() const;
+    // virtual void Save( int currEnergy ) const;
 
   private:
-    void PrepareVolumeOutput() override;
-    void WriteVolumeOutput( unsigned idx_pseudoTime ) override;
+    void GenerateEnergyGrid( bool refinement );
+
+    // IO
+    void PrepareVolumeOutput() override final;
+    void WriteVolumeOutput( unsigned idx_pseudoTime ) override final;
+
+    // Solver
+    void FVMUpdate( unsigned idx_energy ) override final;
+    void FluxUpdate() override final;
+    void IterPreprocessing( unsigned idx_pseudotime ) override final;
+    void virtual IterPostprocessing( unsigned idx_pseudotime ) override final;
+    void SolverPreprocessing() override final;
 };
 
 #endif    // CSDSNSOLVER_H
