+++ src/solvers/pnsolver.cpp
@@ -9,6 +9,7 @@
 #include "toolboxes/textprocessingtoolbox.hpp"
 // externals
 #include "spdlog/spdlog.h"
+
 #include <iostream>
 #include <mpi.h>
